i3c: aspeed: Remove the stop condition after target mode enable
When I3C is in target mode, after enabling the controller, the stop
condition will cause the hardware to automatically generate a Hot-join
request when bit 27 of 0x0 is set to 0 (Auto HJ disabled). This behavior
may cause potential issues because the stop condition is generated by the
software mode, but the Hot-join is a hardware behavior. If the software
mode is disabled too slowly, the SCL from the master may be missed on our
target side, causing unexpected errors.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I8e0575862feb91d5bea1457a98329343389c20af